Fixing of shared preference accounts and addition of navigation drawer item
There was a bug that was found from the previous commit where the first account cannot be created. This bug was fixed but in order to do so, the ability of making a second account was removed for now. A new item was added to the navigation drawer which will act as the switch account button. Work in progress.