Toolbar and navigation drawer fix and shared preference issue
The toolbar now appears and the navigation drawer works fine. There are now two items in the navigation drawer, general information and logout which goes to their respective activities. There was a slight issue with shared preference that was noticed where going to different activities and coming back causes the information to be displayed as null. Work in progress.